Update the backend to MVC structure
